fix mapping priority
